# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Recruitment API",
        "description": "API documentation for Recruitment system",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/auth/forgot-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Quên mật khẩu",
                "description": "API gửi email reset mật khẩu",
                "operationId": "1c662981cefd33a817b78ed144f44559",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "mail_address"
                                ],
                                "properties": {
                                    "mail_address": {
                                        "description": "Email đăng ký",
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Gửi email thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Email không tồn tại"
                    }
                }
            }
        },
        "/api/auth/reset-password/{token}": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Đặt lại mật khẩu",
                "description": "API đặt lại mật khẩu bằng token từ email",
                "operationId": "1e3454614ce388eef7157dab62378363",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Token reset password từ email",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "abc123xyz"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "mail_address",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "mail_address": {
                                        "description": "Email đăng ký",
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@gmail.com"
                                    },
                                    "password": {
                                        "description": "Mật khẩu mới",
                                        "type": "string",
                                        "format": "password",
                                        "example": "NewPass123@"
                                    },
                                    "password_confirmation": {
                                        "description": "Xác nhận mật khẩu mới",
                                        "type": "string",
                                        "format": "password",
                                        "example": "NewPass123@"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đặt lại mật khẩu thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Token không hợp lệ hoặc đã hết hạn"
                    }
                }
            }
        },
...
        "/api/home/company": {
            "get": {
                "tags": [
                    "Home - Company"
                ],
                "summary": "Danh sách công ty",
                "description": "list",
                "operationId": "890ba8ee11d025cbf123c7da8dab9606",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "orders",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filters",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    }
                }
            }
        },
        "/api/home/company/{id}": {
            "get": {
                "tags": [
                    "Home - Company"
                ],
                "summary": "Chi tiết công ty",
                "description": "Method detail",
                "operationId": "58116089e3334cac8c19b41a76fb700b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "[explicite description]",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            }
        },
...
    },
...
}